move executable_path module to the top-level
