+++ src/lib.rs
@@ -14,6 +14,7 @@ extern crate memoffset;
 
 pub mod cli;
 pub mod context;
+mod executable_path;
 mod recorder;
 mod test_checker;
 mod test_spec;
