+++ src/recorder/mod.rs
@@ -1,9 +1,9 @@
 pub mod hole_recorder;
 mod result;
 
+use crate::executable_path::is_unmocked_command;
 use crate::test_spec::command::Command;
 use crate::test_spec::command_matcher::CommandMatcher;
-use crate::test_spec::executable_path::is_unmocked_command;
 use crate::test_spec::{Step, Test};
 use crate::tracer::stdio_redirecting::Redirector;
 use crate::tracer::SyscallMock;
